refactor(fuse): hoist shared async crypto helpers to content_ops.rs
- Create crates/fuse/src/content_ops.rs with fetch_and_decrypt_content_async
  and publish_file_metadata under cfg any(fuse,winfsp)
- A2 scope narrowing: fetch_and_decrypt_file_content stays in each operations.rs
  because macOS uses 3s private NETWORK_TIMEOUT while Windows uses crate::
  block_with_timeout (10s); different timeouts are intentional for sync FUSE path
- operations.rs (macOS) and platform/windows/operations.rs both re-export the two
  shared async helpers from crate::content_ops
- Add pub mod content_ops cfg-gated any(fuse,winfsp) to lib.rs
- cargo test -p cipherbox-fuse: 64 passed
- Windows winfsp side verified by inspection and CI (cannot compile on macOS)
