+++ crates/fuse/src/platform/windows/operations.rs
@@ -260,144 +260,13 @@ pub mod implementation {
         })
     }
 
-    /// Async version of content download + decrypt for background prefetch tasks.
-    pub async fn fetch_and_decrypt_content_async(
-        api: &cipherbox_api_client::ApiClient,
-        cid: &str,
-        encrypted_file_key_hex: &str,
-        iv_hex: &str,
-        encryption_mode: &str,
-        private_key: &[u8],
-    ) -> Result<Vec<u8>, String> {
-        let encrypted_bytes = cipherbox_api_client::ipfs::fetch_content(api, cid)
-            .await
-            .map_err(|e| e.to_string())?;
-        let encrypted_file_key =
-            hex::decode(encrypted_file_key_hex).map_err(|_| "Invalid file key hex".to_string())?;
-        // unwrap_key returns Zeroizing<Vec<u8>> (S3/D-05).
-        let file_key = cipherbox_crypto::ecies::unwrap_key(&encrypted_file_key, private_key)
-            .map_err(|e| format!("File key unwrap failed: {}", e))?;
-        let file_key_arr = zeroizing_32_from_slice(file_key.as_slice(), "Invalid file key length")?;
-
-        let plaintext = if encryption_mode == "CTR" {
-            let iv = hex::decode(iv_hex).map_err(|_| "Invalid file IV hex".to_string())?;
-            let iv_arr: [u8; 16] = iv
-                .try_into()
-                .map_err(|_| "Invalid CTR IV length (expected 16)".to_string())?;
-            cipherbox_crypto::aes_ctr::decrypt_aes_ctr(&encrypted_bytes, &file_key_arr, &iv_arr)
-                .map_err(|e| format!("CTR decryption failed: {}", e))?
-        } else {
-            let iv = hex::decode(iv_hex).map_err(|_| "Invalid file IV hex".to_string())?;
-            let iv_arr: [u8; 12] = iv
-                .try_into()
-                .map_err(|_| "Invalid GCM IV length (expected 12)".to_string())?;
-            cipherbox_crypto::aes::decrypt_aes_gcm(&encrypted_bytes, &file_key_arr, &iv_arr)
-                .map_err(|e| format!("GCM decryption failed: {}", e))?
-        };
-
-        Ok(plaintext)
-    }
-
-    /// Encrypt and publish per-file FileMetadata to the file's own IPNS record.
-    pub async fn publish_file_metadata(
-        api: &cipherbox_api_client::ApiClient,
-        file_meta: &cipherbox_core::folder::FileMetadata,
-        folder_key: &[u8],
-        file_ipns_private_key: &zeroize::Zeroizing<Vec<u8>>,
-        file_ipns_name: &str,
-        coordinator: &crate::PublishCoordinator,
-        tee_public_key: Option<&[u8]>,
-        tee_key_epoch: Option<u32>,
-        is_first_publish: bool,
-    ) -> Result<(), String> {
-        let folder_key_arr = zeroizing_32_from_slice(
-            folder_key,
-            "Invalid folder key length for FileMetadata encryption",
-        )?;
-
-        // Encrypt FileMetadata with parent folder key
-        let sealed = cipherbox_core::folder::encrypt_file_metadata(file_meta, &folder_key_arr)
-            .map_err(|e| format!("FileMetadata encryption failed: {}", e))?;
-
-        // Package as JSON envelope: { "iv": hex, "data": base64 }
-        let iv_hex = hex::encode(&sealed[..12]);
-        use base64::Engine;
-        let data_base64 = base64::engine::general_purpose::STANDARD.encode(&sealed[12..]);
-        let json = serde_json::json!({ "iv": iv_hex, "data": data_base64 });
-        let json_bytes = serde_json::to_vec(&json)
-            .map_err(|e| format!("FileMetadata JSON serialization failed: {}", e))?;
-
-        // Upload encrypted file metadata to IPFS
-        let file_meta_cid = cipherbox_api_client::ipfs::upload_content(api, &json_bytes)
-            .await
-            .map_err(|e| e.to_string())?;
-
-        // Resolve current IPNS sequence number
-        let current_seq = if is_first_publish {
-            None
-        } else {
-            Some(coordinator.resolve_sequence(api, file_ipns_name).await?)
-        };
-
-        // Create and sign IPNS record
-        let ipns_key_arr = zeroizing_32_from_slice(
-            file_ipns_private_key.as_slice(),
-            "Invalid file IPNS private key length",
-        )?;
-        let new_seq = crate::next_file_publish_sequence(is_first_publish, current_seq)?;
-        let value = format!("/ipfs/{}", file_meta_cid);
-        let record =
-            cipherbox_core::ipns::create_ipns_record(&ipns_key_arr, &value, new_seq, 86_400_000)
-                .map_err(|e| format!("File IPNS record creation failed: {}", e))?;
-        let marshaled = cipherbox_core::ipns::marshal_ipns_record(&record)
-            .map_err(|e| format!("File IPNS record marshal failed: {}", e))?;
-
-        let record_b64 = base64::engine::general_purpose::STANDARD.encode(&marshaled);
-
-        // TEE enrollment on first publish only (same pattern as folder creation in write_ops.rs)
-        let (encrypted_ipns_for_tee, tee_epoch) =
-            match (is_first_publish, tee_public_key, tee_key_epoch) {
-                (true, Some(tee_key), Some(epoch)) => {
-                    let wrapped =
-                        cipherbox_crypto::wrap_key(file_ipns_private_key.as_slice(), tee_key)
-                            .map_err(|e| format!("TEE key wrapping failed: {}", e))?;
-                    (Some(hex::encode(&wrapped)), Some(epoch))
-                }
-                (true, Some(_), None) => {
-                    return Err("TEE public key present but key_epoch missing".to_string());
-                }
-                _ => (None, None),
-            };
-
-        // Per-file IPNS publishes do not use conflict detection -- file metadata
-        // is owned by the file's own IPNS keypair and conflicts are inherently
-        // avoided by the per-file sequence number management.
-        let req = cipherbox_api_client::IpnsPublishRequest {
-            ipns_name: file_ipns_name.to_string(),
-            record: record_b64,
-            metadata_cid: file_meta_cid.clone(),
-            encrypted_ipns_private_key: encrypted_ipns_for_tee,
-            key_epoch: tee_epoch,
-            expected_sequence_number: None,
-        };
-        match cipherbox_api_client::ipns::publish_ipns(api, &req)
-            .await
-            .map_err(|e| e.to_string())?
-        {
-            cipherbox_api_client::PublishResult::Success => {}
-            cipherbox_api_client::PublishResult::Conflict { .. } => {
-                log::warn!(
-                    "Unexpected conflict on per-file IPNS publish for {}",
-                    file_ipns_name
-                );
-            }
-        }
-
-        coordinator.record_publish(file_ipns_name, new_seq);
-        log::info!("Per-file IPNS publish succeeded for {}", file_ipns_name);
-
-        Ok(())
-    }
+    // Re-export shared async helpers from content_ops (Tier-2 dedup, Plan 55-03).
+    // fetch_and_decrypt_content_async and publish_file_metadata are identical
+    // between macOS and Windows; they are now in the shared content_ops module.
+    // fetch_and_decrypt_file_content (sync wrapper, uses crate::block_with_timeout
+    // with 10s timeout) stays defined locally here; the macOS path uses its own
+    // 3s timeout copy in operations.rs (A2 scope narrowing).
+    pub use crate::content_ops::{fetch_and_decrypt_content_async, publish_file_metadata};
 
     // ── FileSystemContext Implementation (delegates to sub-modules) ──────
